+++ src/data_object_internal.h
@@ -14,7 +14,7 @@ struct data_object_funcs {
 	int (*read)(data_object_t *obj, void **buf, off_t start, size_t len);
 	int (*write)(data_object_t *obj, off_t offset, void *data, size_t len);
 	int (*free)(data_object_t *obj);
-	int (*get_size)(data_object_t *obj, size_t *size);
+	int (*get_size)(data_object_t *obj, off_t *size);
 };
 
 int data_object_create_impl(data_object_t **obj, void *impl,
